+++ src/utils/react-router-utils.tsx
@@ -40,7 +40,3 @@ export function ReactQueryWrapper(props: React.PropsWithChildren) {
     </QueryClientProvider>
   );
 }
-
-export function cls(...args: unknown[]): string {
-  return args.filter(Boolean).join(" ");
-}
